+++ CHANGELOG.md
@@ -26,6 +26,7 @@ This changelog also contains important changes in dependencies.
 - (usvg) Coordinates resolving when units are `userSpaceOnUse`.
 - (usvg) Groups regrouping. Caused an incorrect rendering of `clipPath`
   that had `filter` on a child.
+- (usvg) Style attributes resolving on the root `svg` element.
 - (qt-backend) Stroking of LineTo specified as CurveTo.
 - (svgdom) `stroke-miterlimit` attribute parsing.
 - (svgdom) `length` and `number` attribute types parsing.